+++ TechShopSolution.AdminApp/Views/Slide/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@using TechShopSolution.ViewModels.Common;
-@model PagedResult<TechShopSolution.ViewModels.Website.SlideViewModel>
+@model PagedResult<TechShopSolution.ViewModels.Website.Slide.SlideViewModel>
 
 @{
     ViewData["Title"] = "Danh sách Slide";
