+++ TechShopSolution.AdminApp/Views/Slide/Update.cshtml
@@ -1,4 +1,4 @@
-﻿@model TechShopSolution.ViewModels.Website.SlideUpdateRequest
+﻿@model TechShopSolution.ViewModels.Website.Slide.SlideUpdateRequest
 
 @{
     ViewData["Title"] = "Chi tiết Slide";
