+++ TechShopSolution.ApiIntegration/SlideApiClient.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using TechShopSolution.ViewModels.Common;
 using TechShopSolution.ViewModels.Website;
+using TechShopSolution.ViewModels.Website.Slide;
 
 namespace TechShopSolution.ApiIntegration
 {
