+++ TechShopSolution.ViewModels/TechShopSolution.ViewModels.csproj
@@ -13,4 +13,8 @@
     <ProjectReference Include="..\TechShopSolution.Data\TechShopSolution.Data.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Website\Contact\" />
+  </ItemGroup>
+
 </Project>
